pyproject.toml
[project]
name = "minecraftbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-cord>=2.7.1",
    "python-dotenv>=1.2.2",
]
